v0.1-windsurf/backend/requirements.txt
flask==2.3.3
python-pptx==0.6.21
pyautogen==0.4.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
pyyaml==6.0.1
pytest==7.4.2
black==23.7.0
pylint==2.17.5
